bug fixes temp20
