+++ src/app/modules/Elementary/elementary.module.ts
@@ -8,14 +8,15 @@ import { Ntemplate5 } from './Ntemplate5/Ntemplate5.component';
 import { Ntemplate6 } from './Ntemplate6/Ntemplate6.component';
 import { Ntemplate7 } from './Ntemplate7/Ntemplate7.component';
 import { Ntemplate8 } from './Ntemplate8/Ntemplate8.component';
+import { Ntemplate11Component } from './Ntemplate11/Ntemplate11.component';
 import { Ntemplate13 } from './Ntemplate13/Ntemplate13.component';
 import { Ntemplate1Component } from './Ntemplate1/Ntemplate1.component';
 import { Ntemplate9Component } from './Ntemplate9/Ntemplate9.component';
 import { Ntemplate18 } from './Ntemplate18/Ntemplate18.component';
-import { Ntemplate24_1 } from './Ntemplate24_1/Ntemplate24_1.component'
+import { Ntemplate24_1 } from './Ntemplate24_1/Ntemplate24_1.component';
 import { Ntemplate20Component } from './Ntemplate20/Ntemplate20.component';
 import { Ntemplate19Component } from './Ntemplate19/Ntemplate19.component';
-
+  
 
 
  
@@ -32,13 +33,14 @@ import { Ntemplate19Component } from './Ntemplate19/Ntemplate19.component';
   Ntemplate6,
   Ntemplate7,
   Ntemplate8,
+  Ntemplate11Component,
   Ntemplate13,
   Ntemplate1Component,
   Ntemplate9Component,
   Ntemplate18,
   Ntemplate24_1,
   Ntemplate20Component,
-  Ntemplate19Component
+  Ntemplate19Component,
   ],
   providers: []
 })
